requirements.txt
msal==1.35.1
python-dotenv==1.2.1
urllib3<2
keyring==25.7.0
PyNaCl==1.5.0
